3.1. In-Component Route Guards
Vue gives us many different component lifecycle hooks, like created(), mounted(), updated(), etc. When using Vue Router, we get three more component hooks called (Route Navigation Guards)[https://router.vuejs.org/guide/advanced/navigation-guards.html#in-component-guards]:
  beforeRouteEnter(routeTo, routeFrom, next)
  beforeRouteUpdate(routeTo, routeFrom, next)
  beforeRouteLeave(routeTo, routeFrom, next)
We can define each of these inside our components, just like lifecycle hooks. First, let’s learn about the Parameters:

routeTo - This refers to the route that is about to be navigated to.
routeFrom - This refers to the route that is about to be navigated away from.
next - This is a function that must be called in each of them to resolve the hook, and continue navigation.
Now let’s take a closer look at when each of these is called when it’s defined inside a component.

beforeRouteEnter(routeTo, routeFrom, next)
This is called before the component is created. Since the component has not been created yet, we can’t use the this keyword here.

beforeRouteUpdate(routeTo, routeFrom, next)
This is called when the route changes, but is still using the same component. An example here is when we paginate, and we switch from page to page but still using the same component. It does have access to \“this\”.

beforeRouteLeave(routeTo, routeFrom, next)
This is called when this component is navigated away from. It does have access to \“this\”.

As I mentioned above, each of these methods at some point must call next(). Here’s what you can do with next:

**next()** - Called by itself will continue navigation to routeTo.

**next(false)** - Cancels the navigation.

**next('/')**- Redirects page to the / path.

**next({ name: 'event-list' })** - Redirects to this named path

Any options you might put in a router-link's to property you can send into next() for redirecting navigation.

Notice that there’s no progress bar and the tempate shows up on the page before the data gets returned from the API call. We only want to see the template when the data arrives.

We have three steps to implement our solution:
 1.Start the progress bar when routing to the component.
 2.Make the API call, using the fetchEvent action.
 3.When API returns finish progress bar.
 4.Render the component template.

First, we include NProgress and the Vuex Store. Then we use the beforeRouteEnter navigation guard. This guard is called before the component is created, which means it doesn’t have access to this . That’s why we needed to import the store and why we’re calling the action explicitly and not using mapAction, which I’ve removed.

We start the progress bar, dispatch the fetchEvent Action, and then when the Action is done we finish the progress bar and call next(), which tells Vue to continue navigation to this component. Until next() is called, the web app is going to wait.

If we look in the browser, we see the progress bar working, but it doesn’t seem to be waiting for our API call to be finished before finishing the progress bar. What is going on?