4.2. Global and Per-Route Guards
Use a Route Guard inside our router, rather than inside our component.
By including our Vuex Store and using the Per-Route inside our router.js we can dispatch our fetchEvent Action (that we normally would call from our EventShow), and only once that’s returned will we allow the navigation to continue by calling next(). However, in order for then() to get called when the API is returned we need to make sure this action returns a promise (we did this once before in the previous lesson if you’re following along)
Now that we’re calling the Action from our router, we no longer need to call it from our EventShow action.
And if I look in the browser, everything is working as expected.