5.1. Completing our progress bar
We want our progress bar to run and finish when our API call is returned and the data is loaded into the EventList template.

In order to do this, there are a few things we need to do:

1.Refactor the perPage variable into our state.
2.Modify the router.js file
3.Use beforeRouteEnter and beforeRouteUpdate in the EventList component.

Step 1: We need perPage to be in the state
We needed some way to keep track of how many events to display per page, and previously we used a local constant called perPage in our EventList component.
Declaring this.perPage = 3 in our created() lifecycle hook added perPage as a local constant we could access in other methods (like hasNextPage()).
In a minute we’re going to do what we did in the previous lesson inside EventShow, where we moved the Vuex Action outside of the created() lifecycle hook and into a route guard.
Since the route guards run before the component is instantiated, this does not exist yet. So, we need to move perPage into the Vuex State. You might think it belongs in the State, and we never should have put it in the component to begin with. However, there’s nothing wrong with placing it in either place.
All we changed here is adding perPage to the event State with a default of 3, and then we extracted the state parameter from the context object, and used the State inside our getEvent's call. We also added a return to our EventService so we can only render our EventList component when the API call is finished.

That’s our next step:

Step 2 - Modify the router

At the moment, our router code looks like this:

📃 /src/router.js

    ...
    const router = new Router({
      mode: 'history',
      routes: [
        {
          path: '/',
          name: 'event-list',
          component: EventList
        },
        ...
Here we want to dispatch the fetchEvents Action before the component is loaded like we did with EventShow with beforeEnter.
With all we’ve learned, you might think the code below would be what we need. However, 🛑 This won’t work! Take a look at this code below, and then we’ll learn why it won’t work
 /src/router.js

    ...
    const router = new Router({
      mode: 'history',
      routes: [
        {
          path: '/',
          name: 'event-list',
          component: EventList,
          beforeEnter(routeTo, routeFrom, next) {
            const currentPage = parseInt(routeTo.query.page || 1)
            store
              .dispatch('event/fetchEvents', {
                page: currentPage
              })
              .then(() => {
                ...
If I were to use this code to call fetchEvents it would only get called when the route is initially loaded. Then when we hit the next page link, nothing would happen. ****Our route gets updated when we click on the next page, but since the component is not completely reloaded, fetchEvents doesn’t get called.
So we need another way to ensure fetchEvents is called both when the page is navigated to, and when we paginate.

For now, let’s finish up the code we need in router.js by setting props: true because we’ll be setting our current page as a parameter to pass in as a prop.

Step 3 - Use beforeRouteEnter and beforeRouteUpdate in the EventList component

From our lesson on In-Component Route guards we learned that beforeRouteEnter is called when the component is created and beforeRouteUpdate is called when the route changes, but we’re still using the same component. This is what we need to ensure fetchEvents is called when we paginate.