5.3. Completing our progress bar
Start the progress bar when creating an event. When redirects to show event route after successful event creation the progress bar is already started so it's done after displaying the created event by before written code.
Also we stop the progress bar if creation of the event fails for whatever reasons.