9.3. Reusable Form Components: BaseButton
Style the button as we want

Dynamic Button Styles
Now we need to think about how we can add dynamic styles to our button. You might think that the solution would be to set inheritAttrs to false like we did in our BaseInput lesson, which allowed us to inherit attributes added on the parent scope. So it would make sense that we could just bind those attributes right onto the native button element.
However, this isn’t currently how Vue works. As of now, the class and style attributes aren’t available to you on $attrs. However, we can expect that to change in Vue 3.
So if we can’t v-bind to our $attrs, how can we apply a dynamic class to our BaseButton? We can use props for that.
Now we have a buttonClass prop that we’ve bound to the button. This means we can now pass through class names from the parent scope, like our class of -fill-gradient.