9.4. Reusable Form Components: BaseButton
Even though class and style aren’t currently available to us in $attrs, we’ll still want to be inheriting attributes because we still want to have the ability to inherit attributes added on the parent scope, such as disabled.

For example:

  <BaseButton type="submit" buttonClass="-fill-gradient" disabled>Submit</BaseButton>
So we’ll add the ability to inherit attributes to BaseButton now.