10.4. Form Validation with VueValidate
Only yield messages to user when user interacts with email field. At first loading there's no messages displayed.

We can give the user a chance to fill it out before telling them it’s required. In other words, if a user touches the form then clicks away, we can let them know that it’s required.
Fortunately, there’s a name for that: When a user focuses on an element and then focuses away from it. It’s called blur. So we can listen for the blur event on our input.
So the question now is: What do we want to trigger when the blur event happens?
The way Vuelidate works, validations are run on each change, as we observed earlier. But in order for the $error state to be true, a field needs to be both $dirty and $invalid. So we need to be triggering the $dirty state to be true after a user has interacted with the field so that we can make this field have an $error. We then can use that $error state to our advantage, as we’ll see in a minute.
But what does it mean for a field to be $dirty? That term just means the field has been interacted with, or in other words: touched. We can imagine that once it’s been clicked on or input into, that it’s now dirty with the fingerprint of our user. So in order to change the state of this field to $dirty, we can trigger Vuelidate’s touch method on blur.
Now once a user has focused on this input, then focused away from it, the $dirty state will be true, and if the field is also blank, then $v.email.required will be false, which means the field is $invalid. Again, the way Vuelidate works, if a field is both $dirty and $invalid then that means there is an $error.
So now we can use this $error state to our advantage. We can wrap our error messages in a div, and only display that div once it is true that there is an $error on this field.
Great. Now when this component first loads, our error messages will be hidden until an $error occurs, which means the user has actually touched the field and left it blank, or input an incorrect value.