11.2. Form Validation with VueValidate Pt2
Validate category property of the event. Currently within our BaseSelect component that native select element isn't able to hear the listeners that we added to parent scope. So isn't able to hear the @blur event we added at the parent scope so when the category field get the focus and leave the focus unselected no error message is displayed because $touched method isn't fired and $dirty property on $v.event.category is not set.  $v object is built in vuelidate plugin