11.5. Form Validation with VueValidate Pt2
UI for title event field. We have the same problem as in 11.2 commit that have the same solution but we run in another problem. When we try to type into title field we get a weird [object InputEvent] not what we typed into.
Great it’s working so far. But when we actually try to input a title here, we get this weird [object InputEvent] showing up.
Oh geez… what’s going on here? The reason for this strange behavior is a bit complicated to grasp, so let’s tackle this piece by piece.

We have v-model on our BaseInput:

📃 src/views/EventCreate.vue

    <BaseInput
      label="Title"
      placeholder="Title"
      type="text"
      class="field"
      :class="{ error: $v.event.title.$error }"
      @blur="$v.event.title.$touch()"
    />
As we covered in our lesson where we built BaseInput, v-model is really just syntactic sugar. In this case, v-model="event.title" translates to:

    :value="event.title"
    @input="(value) => { event.title = value }"
Why is this important? Well, this means we have a listener here for the input event. And since we just added v-on="$listeners" that means we are inheriting that input event listener onto our BaseInput’s input element.

    <input :value="value" v-bind="$attrs" @input="updateValue" v-on="$listeners">
And why is that important? Well, look again at what’s on that element. We’ve also got @input="updateValue". That means we have two input event listeners. One that is triggering the updateValue method, and one that we’re inheriting from the parent scope, which is hidden under the sugar-coating of our v-model. This means we have a conflict, and the second input (the one from the v-model) is taking precedence.